Cleanup loadImage function
